requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.2
requests==2.31.0
liac-arff==2.5.0
joblib==1.3.2
pydantic==2.5.0